Update set-theory.html.markdown
Rewrote the outline of the previous article.
I really like set theory so I wanted to make this article good. I wrote this in about 90 minutes and spell checked it.

I hope it is not too long. I apologize for the large expository elements, since set theory and rigorous mathematical background is something I cannot assume from intermediate programmers, usually. Generally, the article is organized so that the upper part contains the quick what-you-need-to-know about set theory, and the second part is going more into the difference between naive and axiomatic set theory. This is much more difficult to show examples of, especially the axiom of choice, but most people doing anything requiring sets will have to deal with the axiom of choice, so I should at least mention it.